Move audio focus method into their own class for efficiency
